QA/requirements-test.txt
# Test Requirements and Dependencies

# Core Testing Framework
pytest==7.4.3
pytest-cov==4.1.0
pytest-flask==1.3.0
pytest-mock==3.12.0
pytest-asyncio==0.21.1
pytest-benchmark==4.0.0
pytest-xdist==3.5.0  # Parallel test execution

# Test Data Generation
factory-boy==3.3.0
faker==20.1.0

# API Testing
requests==2.31.0
requests-mock==1.11.0

# Database Testing
pytest-postgresql==5.0.0
pytest-alembic==0.10.7

# Coverage and Reporting
coverage[toml]==7.3.3
pytest-html==4.1.1
pytest-json-report==1.5.0

# Performance Testing
locust==2.19.1

# Security Testing
bandit==1.7.5
safety==2.3.5

# Code Quality
flake8==6.1.0
pylint==3.0.3
black==23.12.1
mypy==1.7.1

# Mocking and Stubs
responses==0.24.1
freezegun==1.4.0  # Time mocking

# Image Testing
opencv-python==4.8.1.78
Pillow==10.1.0

# Additional Utilities
python-dotenv==1.0.0
colorama==0.4.6  # Colored output
